Split to different file components and cleanup
