+++ client/src/components/common/Header/ShoppingBag/shopping-bag.js
@@ -5,8 +5,9 @@ import styled from "styled-components";
 import bug from "./shopping-bag.png";
 import { mediaMobile } from "../../../../styled-components/media-breakpoints-mixin";
 export const ShoppingBag = () => {
-  // Use state quantitybyid if not logged in,
-  // after login use state.shoppingCart.serverProducts
+  // Use state Local Store Cart Quantity if not logged in,
+  // after login use state.shoppingCart.srvCart Cart Quantity from Server
+  // After Login the Two Carts are merged
   const count = useSelector(state => {
     const localCartCount = Object.values(state.shoppingCart.locCart).reduce(
       (acc, curr) => acc + curr,
